+++ backend/src/main/java/com/metamong/mt/domain/member/model/constant/IdOrPw.java
@@ -1,0 +1,6 @@
+package com.metamong.mt.domain.member.model.constant;
+
+public enum IdOrPw {
+    ID,
+    PW
+}
