+++ backend/src/main/java/com/metamong/mt/global/config/BeanConfig.java
@@ -12,8 +12,8 @@ import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.databind.module.SimpleModule;
 import com.fasterxml.jackson.databind.util.StdDateFormat;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.metamong.mt.domain.member.model.IdOrPw;
-import com.metamong.mt.global.jackson.role.IdOrPwDeserializer;
+import com.metamong.mt.domain.member.model.constant.IdOrPw;
+import com.metamong.mt.global.jackson.idorpw.IdOrPwDeserializer;
 import com.metamong.mt.global.mail.MailAgent;
 import com.metamong.mt.global.mail.MailAgentMock;
 import com.metamong.mt.global.mail.MailMessageFormatter;
