+++ backend/src/main/java/com/metamong/mt/global/jackson/idorpw/IdOrPwDeserializer.java
@@ -1,4 +1,4 @@
-package com.metamong.mt.global.jackson.role;
+package com.metamong.mt.global.jackson.idorpw;
 
 import java.io.IOException;
 
@@ -6,7 +6,7 @@ import com.fasterxml.jackson.core.JacksonException;
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonDeserializer;
-import com.metamong.mt.domain.member.model.IdOrPw;
+import com.metamong.mt.domain.member.model.constant.IdOrPw;
 
 public class IdOrPwDeserializer extends JsonDeserializer<IdOrPw> {
